[DEV-21440] Initial commit implementing jira_sync
Tests are incoming

This repo is going to contain an implementation of a Danger plugin that syncs data between Jira and GitHub. It currently includes one command, `autolabel_pull_request`, which will label Pull Requests with all component names and project keys in related Jira  issues.

See `README.md` for usage and details