Fixes bug with adding GitHub labels that preexist
Previously, tests expected that `plugin.github.api.labels` would return a list of strings representing the names of labels. It turns out that each label encompasses much more data than that, and we were comparing hashes to strings when trying to avoid re-adding existing labels.

This commit changes the mocked response to `plugin.github.api.labels` to more closely match that of reality.

This commit also fixes a broken test that was passing but not testing what I thought it was.